sb: more proper usage of P4-Context
As spotted by Elliott Sales de Andrade, the P4-Context alias is not
cleared.

This is probably because when there's no private-alias, alias_backup is
NULL, and therefore no restoration happens. For starters public-alias
should be safer because essentially everyone has one so alias_backup
will never be NULL, but the proper way to ensuer is with an independent
flag.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>