Reorganize peer code for tracking sbs better
We a sb is closed we want to close all the calls, and their respective
peer msgs, not the links. Therefore sb and link are now decoupled, and
call is used instead.

Also, in order to be able to send messages the originator sb must
be saved in the peer msg.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>