sb: close on inactivity
This fixes the long-standing problem of undelivered messages due to
switchboard timeouts.

In 1 to 1 conversations in normal conditions (when chatting with a
contact that uses WLM official client in Windows) the contact leaves
after 1 minute of inactivity; We would receive a BYE command, and
immediately afterward the connection should be closed. Other clients
will not leave the sb, in such cases the server will close it after 5
minutes.

However, in certain network conditions the BYE command is never
received, nor the notification of the closed connection. In such cases we
think the sb is alive so we send commands, but we never get
acknowledgement because the connection is already dead.

Further investigation would be required in order to determine why we are
not notified of the disconnection, but in order to workaround the issue
we now close sb after 1 minute of inactivity.

This is possibly not the best thing to do, but it emulates what the
official client does and for user's point of view it's perfect.

Fixes issue #181.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>