Fix authentication
Apparently Microsoft turned off the lod nexus server for authentication.
So now we are forced to use the Passport server.

Fortunately, we already have the code to authenticate to the Passport
server on the OIM code. So we just need to copy it.

We would need to reorganize this code to reuse the tickets all over.

This gets rid of the "nexus stream error".

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>